Filter dataset with blacklist of genotypes

Filter dataset with blacklist of genotypes.

This function allows to blacklist/erase/mask genotypes.

Filter target: Individual genotype calls.

Used internally in radr and might be of interest for users.

Usage

filter_blacklist_genotypes(data, blacklist.genotypes, verbose = TRUE, ...)

Arguments

data

(4 options) A file or object generated by radr:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Use read_genome to import supported formats and tidy_genome when a tidy table is needed.

blacklist.genotypes

A required blacklist object or filepath containing genotype records to remove or mask.

verbose

Logical indicating whether progress messages are emitted. Default: verbose = TRUE.

...

Additional arguments passed to lower-level screening or filtering functions.

Life cycle

This function arguments will be subject to changes. Currently the function uses erase.genotypes, but using the dots-dots-dots ... arguments allows to pass erase.genotypes and masked.genotypes. These arguments do exactly the same thing and only one can be used.

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) { # \dontrun{
data <- radr::filter_blacklist_genotypes(
    data = data, blacklist.geno = "blacklist.geno.tsv"
    )
} # }