Filter common markers between strata

The function will filter the markers by keeping only those in common between all strata (population or any groupings defined in STRATA column).

Filter targets: SNPs

Statistics: strata genotyping rate per SNPs

Used internally in radr and might be of interest for users who wants to keep only markers in common.

Usage

filter_common_markers(
  data,
  filter.common.markers = TRUE,
  fig = FALSE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)

Arguments

data

(4 options) A file or object generated by radr:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Use read_genome to import supported formats and tidy_genome when a tidy table is needed.

filter.common.markers

(optional, logical) Default: filter.common.markers = TRUE.

fig

(optional, logical) fig = TRUE will produce a ComplexUpset figure to visualize the number of markers between populations. Default: fig = FALSE.

parallel.core

Number of workers available for parallel operations. Default: parallel.core = parallel::detectCores() - 1.

verbose

(optional, logical) verbose = TRUE to be chatty during execution. Default: verbose = FALSE.

...

(optional) To pass further arguments for fine-tuning the function and legacy arguments.

Value

A list with the filtered input, whitelist and blacklist of markers..

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) { # \dontrun{
require(SeqArray) # when using gds
common <- radr::filter_common_markers(data = "my.radr.gds.rad", verbose = TRUE)
} # }